# API Layer/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Smart App Group C - API Layer
  description: |-
    This is the API used for internal communication between services and by external applications and tools to allow data access.
    Made by the integration team: Andrea Alberti, Anna Francesca Montagnoli, Matteo Simone
  version: 1.1.1
servers:
- url: /
...
paths:
...
  /user/login:
    post:
      tags:
      - user
      summary: Logs user into the system
      description: "Attempts to log a user into the system given with their credentials.\
        \ On success, returns a token that is needed in the 'Authorization' header\
        \ of protected API calls. The token has a TTL."
      operationId: user_login_post
      parameters:
      - name: username
        in: header
        description: The username for login
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: password
        in: header
        description: The password for login in clear text
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationInfo'
        "401":
          description: Wrong credentials error
          content:
            application/json:
              schema:
                type: string
                example: Wrong credentials
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/logout:
    get:
      tags:
      - user
      summary: Logs out current user
      description: "Logs out current user, invalidating their authentication token."
      operationId: user_logout_get
      responses:
        "200":
          description: Success Message
          content:
            application/json:
              schema:
                type: string
                example: Success
                x-content-type: application/json
        "401":
          description: Access token not valid
          content:
            application/json:
              schema:
                type: string
                example: Unauthorized
                x-content-type: application/json
        "500":
          description: Can't connect to the requested service
          content:
            application/json:
              schema:
                type: string
                example: Connection Error
                x-content-type: application/json
      security:
      - api_token_auth: []
      x-openapi-router-controller: swagger_server.controllers.user_controller
...
components:
  schemas:
...
    AuthenticationInfo:
      type: string
      description: The authentication token to be used in subsequent API calls
      example: Bearer axi9enflrn
...
  securitySchemes:
    api_token_auth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_api_token_auth
    machine_auth:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://SA2024C.com/oauth/token
          scopes:
            read:machine: Read machine data
            write:machine: Write machine data
            read:KPI: Read KPI data
            write:KPI: Write KPI data
            read:user: Access user data
            write:user: Modify user data
            read:alert: Read alert data
            write:alert: Edit alert data
            read:report: Read SMO report data
            write:report: Edit report data
            read:aiquery: Get ai query responses
            write:aiquery: Request new ai queries
            read:data: Access raw data
        authorizationCode:
          authorizationUrl: https://SA2024C.com/oauth/authorize
          tokenUrl: https://SA2024C.com/oauth/token
          scopes:
            read:machine: Read machine data
            write:machine: Write machine data
            read:KPI: Read KPI data
            write:KPI: Write KPI data
            read:user: Access user data
            write:user: Modify user data
            read:alert: Read alert data
            write:alert: Edit alert data
            read:report: Read SMO report data
            write:report: Edit report data
            read:data: Access raw data
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_machine_auth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_machine_auth